# spec/openapi.yaml
openapi: 3.1.0
info:
  title: 국장레이더 API
  version: 1.0.0
  description: |
    뉴스와 국내 상장 종목의 연결을 탐색하는 API.
    모든 점수는 '연결 강도'이며 미래 수익률 예측이 아니다.
    타입 정의 원본은 spec/types.ts.
servers:
  - url: https://api.gukjang-radar.kr
  - url: http://localhost:3000/api
...
paths:
...
  /v1/auth/me:
    get:
      tags: [auth]
      security: [ { bearerAuth: [] } ]
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/AppUser' } } }
        '401': { description: 미로그인 }
    delete:
      tags: [auth]
      summary: 회원 탈퇴 — alert_keyword/push_subscription은 cascade 삭제(개인정보처리방침 §5)
      security: [ { bearerAuth: [] } ]
      responses:
        '204': { description: 탈퇴 완료 }
        '401': { description: 미로그인 }
...
  /v1/alerts/{alertId}:
    delete:
      tags: [alerts]
      security: [ { bearerAuth: [] } ]
      parameters: [ { name: alertId, in: path, required: true, schema: { type: integer } } ]
      responses: { '204': { description: 삭제됨 } }
  /v1/push/subscribe:
    post:
      tags: [alerts]
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoint, keys]
              properties:
                endpoint: { type: string }
                keys:
                  type: object
                  required: [p256dh, auth]
                  properties:
                    p256dh: { type: string }
                    auth: { type: string }
      responses: { '201': { description: 구독 등록 } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
  schemas:
...
    AppUser:
      type: object
      required: [id, email, provider, plan, createdAt]
      properties:
        id: { type: integer }
        email: { type: [string, 'null'] }
        provider: { type: [string, 'null'], enum: [kakao, google, null] }
        plan: { type: string, enum: [FREE, PRO, PRO_PLUS], description: V1은 항상 FREE(PRD D1) }
        createdAt: { type: string, format: date-time }
...